requirements.txt
python-dotenv
discord.py
openai
better-profanity